Revamp the extras file to find the correct version.
The current yaml_cpp_vendor-extras.cmake file first
attempts to find the yaml-cpp package, and only if
that fails goes ahead and adds the vendored package
to the path.

But that means that currently, if we force build a
vendor package, downstream packages will still find the
system one if both are installed.  That isn't the desired
behavior.

Revamp this file so that we unconditionally add in the
location where the vendor package would be first, then
try to find it.  If that location exists, and has the
vendored package, we'll use that.  If that location
does not exist, we'll still go searching in the normal
OS directories for the package, which should find the
system one (if it exists).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>